Logical data type & Logical operations
• logical statements revised/debugged
* • (sigproc internal) 1) name change (datachunk to body), 2) datatype
logical—this is possible for both audio and non-audio (a logical array
is the result of a logical operation of a non-logical array, so it can
be either audio or non-audio), 3) bufBlockSize is the number of bytes of
each data point (i.e., 8 for double, 16 for complex, 1 for string and
logical), 4) audio_var(logical array) is a cascade audio array (the
signal is non-null only if the condition is met), nonaudio_var(logical)
is an extracted array with the indices meeting the condition (the same
as MATLAB). 5) New grammar: var(scalar) = array means insert array at
scalar-th position for non-audio, scalar-ms for audio. The following
means replace: audiovar(t1~t2)=signal, audiovar(id1:id2)=signal,
nonaudio(id1:id2)=vector, duration of left and right hand sides can be
different (meaning that if so the duration would change after the
statement).

+++ Auxtra.cpp
@@ -348,7 +348,7 @@ int GetPropVal(CAstSig &ast, const AstNode *pnode, CGobj *hobj, string propname,
 	}
 	else if (propname=="children")
 	{
-		out.UpdateBuffer(hobj->child.size());
+		out.UpdateBuffer((int)hobj->child.size());
 		for (size_t k=0; k<hobj->child.size(); k++)
 			out.buf[k] = (double)(int)hobj->child[k];
 	}
@@ -490,7 +490,7 @@ int GetPropVal(CAstSig &ast, const AstNode *pnode, CGobj *hobj, string propname,
 		if (hobj->type=='f')
 		{
 			CFigure *cfig = static_cast<CFigure *>(hobj);
-			out.UpdateBuffer(cfig->ax.size());
+			out.UpdateBuffer((int)cfig->ax.size());
 			for (size_t k=0; k<cfig->ax.size(); k++)
 				out.buf[k] = (double)(int)cfig->ax[k];
 		}
@@ -502,7 +502,7 @@ int GetPropVal(CAstSig &ast, const AstNode *pnode, CGobj *hobj, string propname,
 		if (hobj->type=='f')
 		{
 			CFigure *cfig = static_cast<CFigure *>(hobj);
-			out.UpdateBuffer(cfig->text.size());
+			out.UpdateBuffer((int)cfig->text.size());
 			for (size_t k=0; k<cfig->text.size(); k++)
 				out.buf[k] = (double)(int)cfig->text[k];
 		}
@@ -514,7 +514,7 @@ int GetPropVal(CAstSig &ast, const AstNode *pnode, CGobj *hobj, string propname,
 		if (hobj->type=='a')
 		{
 			CAxis *pax = static_cast<CAxis *>(hobj);
-			out.UpdateBuffer(pax->m_ln.size());
+			out.UpdateBuffer((int)pax->m_ln.size());
 			for (size_t k=0; k<pax->m_ln.size(); k++)
 				out.buf[k] = (double)(int)pax->m_ln[k];
 		}
